+++ src/pipecat/services/deepgram.py
@@ -138,6 +138,13 @@ class DeepgramSTTService(STTService):
         merged_options = default_options
         if live_options:
             merged_options = LiveOptions(**{**default_options.to_dict(), **live_options.to_dict()})
+
+        # deepgram connection requires language to be a string
+        if isinstance(merged_options.language, Language) and hasattr(
+            merged_options.language, "value"
+        ):
+            merged_options.language = merged_options.language.value
+
         self._settings = merged_options.to_dict()
 
         self._client = DeepgramClient(